Module westmint_runtime::xcm_config

Structs

Ancestry

AssetsPalletLocation

CheckingAccount

Local

MaxInstructions

ParentOrParentsPlurality

RelayChainOrigin

RelayNetwork

UnitWeightCost

WestendLocation

XcmConfig

Type Definitions

AssetTransactors

Means for transacting assets on this chain.

Barrier

CurrencyTransactor

Means for transacting the native currency on this chain.

FungiblesTransactor

Means for transacting assets besides the native currency on this chain.

LocalOriginToLocation

Local origins on this chain are allowed to dispatch XCM sends/executions.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.