Crate westmint_runtime

Westmint Runtime

Westmint is the testnet for Statemint.

Re-exports

pub use parachains_common as common;

Modules

api

constants

xcm_config

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetAccountDeposit

AssetDeposit

AssetsStringLimit

DepositBase

DepositFactor

ExistentialDeposit

ForeignAssetsApprovalDeposit

ForeignAssetsAssetAccountDeposit

ForeignAssetsAssetDeposit

ForeignAssetsAssetsStringLimit

ForeignAssetsMetadataDepositBase

ForeignAssetsMetadataDepositPerByte

GenesisConfig

MaxCandidates

MaxInvulnerables

MaxPending

MaxProxies

MaxSignatories

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinCandidates

NftsAttributeDepositBase

NftsCollectionDeposit

NftsDepositPerByte

NftsItemDeposit

NftsMaxDeadlineDuration

NftsMetadataDepositBase

NftsPalletFeatures

Offset

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

PotId

ProxyDepositBase

ProxyDepositFactor

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

TransactionByteFee

Relay Chain TransactionByteFee / 10

UniquesAttributeDepositBase

UniquesCollectionDeposit

UniquesDepositPerByte

UniquesItemDeposit

UniquesMetadataDepositBase

Version

Enums

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeEvent

RuntimeFreezeReason

RuntimeHoldReason

RuntimeLockId

RuntimeSlashReason

Constants

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetTxPayment

Assets

AssetsForceOrigin

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

CumulusXcm

DmpQueue

Executive

Executive: handles dispatch to the various modules.

ForeignAssets

ForeignAssetsInstance

Assets managed by some foreign location. Note: we do not declare a ForeignAssetsCall type, as this type is used in proxy definitions. We assume that a foreign location would not want to set an individual, local account as a proxy for the issuance of their assets. This issuance should be managed by the foreign location’s governance.

Migrations

Migrations to apply on runtime upgrade.

Multisig

Nfts

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Proxy

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

TrustBackedAssetsInstance

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

XcmpQueue