Crate westmint_runtime

Westmint runtime.

Re-exports

pub use statemint_common as common;

Modules

api

constants

opaque
Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

Ancestry

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetDeposit

AssetsStringLimit

AttributeDepositBase

ClassDeposit

DepositBase

DepositFactor

DepositPerByte

DisabledValidatorsThreshold

ExistentialDeposit

GenesisConfig

InstanceDeposit

KeyLimit

MaxCandidates

MaxDownwardMessageWeight

MaxInvulnerables

MaxLocks

MaxPending

MaxProxies

MaxReserves

MaxSignatories

MetadataDepositBase
Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinCandidates

MinimumPeriod

Offset

OnRuntimeUpgrade

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParentOrParentsPlurality

Period

PotId

ProxyDepositBase

ProxyDepositFactor

RelayChainOrigin

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionLength

TransactionByteFee
Relay Chain TransactionByteFee / 10

UncleGenerations

UniquesMetadataDepositBase

UniquesStringLimit

UnitWeightCost

ValueLimit

Version

WestendLocation

XcmConfig

Enums

Call

Event

OriginCaller

ProxyType
The type used to represent the kinds of proxying allowed.

Constants

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage
Complex storage builder stuff.

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

Address
The address format for describing accounts.

AllModules	Deprecated
All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem	Deprecated
All modules included in the runtime as a nested tuple of types.

AllPallets
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem
All pallets included in the runtime as a nested tuple of types.

Assets

AssetsForceOrigin

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Barrier

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

CumulusXcm

DmpQueue

Executive
Executive: handles dispatch to the various modules.

LocalAssetTransactor
Means for transacting assets on this chain.

LocalOriginToLocation
No local origins on this chain are allowed to dispatch XCM sends/executions.

LocationToAccountId
Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

Proxy

RandomnessCollectiveFlip

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

XcmOriginToTransactDispatchOrigin
This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter
The means for routing XCM messages which are not for local execution into the right message queues.

XcmpQueue